refactor(Backend): Fix file naming and functions naming breaking the norm.
